--- app/common/countries/DefaultCountryNotFound.java
@@ -1,8 +1,0 @@
-package common.countries;
-
-public class DefaultCountryNotFound extends RuntimeException {
-
-    public DefaultCountryNotFound() {
-        super("No valid country defined in configuration file");
-    }
-}
