--- app/common/countries/InvalidCountryCode.java
@@ -1,8 +1,0 @@
-package common.countries;
-
-public class InvalidCountryCode extends IllegalArgumentException {
-
-    public InvalidCountryCode(String countryCodeAsString) {
-        super("Invalid country code " + countryCodeAsString);
-    }
-}
