--- app/common/models/CommonDataBuilder.java
@@ -1,33 +1,0 @@
-package common.models;
-
-import io.sphere.sdk.categories.Category;
-import io.sphere.sdk.products.ProductProjection;
-
-import java.util.Collections;
-import java.util.List;
-import java.util.Locale;
-
-public class CommonDataBuilder {
-    private UserContext userContext;
-    private Locale fallbackLocale = Locale.ENGLISH;
-    private List<Category> rootCategories;
-    private List<ProductProjection> products = Collections.emptyList();
-
-    private CommonDataBuilder(final UserContext userContext, final List<Category> rootCategories) {
-        this.userContext = userContext;
-        this.rootCategories = rootCategories;
-    }
-
-    public CommonData build() {
-        return new CommonData(userContext, fallbackLocale, rootCategories, products);
-    }
-
-    public CommonDataBuilder products(final List<ProductProjection> products) {
-        this.products = products;
-        return this;
-    }
-
-    public static CommonDataBuilder of(final UserContext userContext, final List<Category> rootCategories) {
-        return new CommonDataBuilder(userContext, rootCategories);
-    }
-}
