--- app/io/sphere/sdk/play/setupwidget/views/success.scala.html
@@ -1,6 +1,0 @@
-<p>The configuration file @io.sphere.sdk.play.setupwidget.controllers.SetupController.PATH has been created.</p>
-
-<a id="reload-link" href="@generalpages.controllers.routes.HomeController.index.url">Reload the application.</a>
-
-
-
