+++ common/app/countries/DefaultCountryNotFound.java
@@ -1,0 +1,8 @@
+package countries;
+
+public class DefaultCountryNotFound extends RuntimeException {
+
+    public DefaultCountryNotFound() {
+        super("No valid country defined in configuration file");
+    }
+}
