+++ common/app/countries/InvalidCountryCode.java
@@ -1,0 +1,8 @@
+package countries;
+
+public class InvalidCountryCode extends IllegalArgumentException {
+
+    public InvalidCountryCode(String countryCodeAsString) {
+        super("Invalid country code " + countryCodeAsString);
+    }
+}
