+++ common/app/io/sphere/sdk/play/setupwidget/models/SphereCredentials.java
