+++ common/app/io/sphere/sdk/play/setupwidget/views/setup.scala.html
