+++ common/app/io/sphere/sdk/play/setupwidget/views/success.scala.html
@@ -1,0 +1,3 @@
+<p>The configuration file @io.sphere.sdk.play.setupwidget.controllers.SetupController.PATH has been created.</p>
+
+<a id="reload-link" href="/">Reload the application.</a>
