+++ common/app/views/main.scala.html
@@ -1,0 +1,32 @@
+@(title: String, data: CommonData)(content: Html)
+
+<!DOCTYPE html>
+
+<html>
+    <head>
+        <title>@title</title>
+        <link rel="stylesheet" media="screen" href="@routes.Assets.at("stylesheets/main.css")">
+        <link rel="shortcut icon" type="image/png" href="@routes.Assets.at("images/favicon.png")">
+    </head>
+    <body>
+        <h3>
+            Notice this template is in a very early stage of development.
+            Please check the <a href="https://github.com/sphereio/sphere-sunrise">project documentation</a> for more information.
+        </h3>
+        
+        <header>
+            <a href="/">@title</a>
+        </header>
+        
+        <section>
+
+            <ol id="categories">
+                @for(category <- data.rootCategories){
+                    <li>@category.getName.get(data.locales()).orElse("")</li>
+                }
+            </ol>
+            @content
+        </section>
+        
+    </body>
+</html>
