+++ common/conf/common.routes
@@ -1,0 +1,5 @@
+
+POST        /setup-credentials        @io.sphere.sdk.play.setupwidget.controllers.SetupController.processForm()
+
+# Map static resources from the /public folder to the /assets URL path
+GET         /assets/*file             controllers.Assets.at(path="/public", file)
