+++ product-catalog/app/products/controllers/ProductController.java
@@ -1,11 +1,20 @@
 package products.controllers;
 
-import play.mvc.Controller;
+import controllers.SunriseController;
+import io.sphere.sdk.categories.CategoryTree;
+import io.sphere.sdk.client.PlayJavaSphereClient;
+import play.Configuration;
 import play.mvc.Result;
 
-public class ProductController extends Controller {
+import static play.mvc.Results.ok;
 
-    public static Result index() {
+public class ProductController extends SunriseController {
+
+    public ProductController(final PlayJavaSphereClient client, final CategoryTree categoryTree, final Configuration configuration) {
+        super(client, categoryTree, configuration);
+    }
+
+    public Result index() {
         return ok("Good!");
     }
 }
